+++ packages/webview_flutter/webview_flutter_android/test/legacy/webview_android_widget_test.mocks.dart
@@ -376,16 +376,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
       ) as _i4.Future<void>);
 
   @override
-  _i4.Future<void> setCacheMode(_i2.CacheMode? mode) => (super.noSuchMethod(
-        Invocation.method(
-          #setCacheMode,
-          [mode],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-
-  @override
   _i4.Future<void> setTextZoom(int? textZoom) => (super.noSuchMethod(
         Invocation.method(
           #setTextZoom,